lib headers
```
/__w/curl/curl/lib/headers.c:57:15: error: conversion to 'unsigned int' from 'int' may change the sign of the result [-Werror=sign-conversion]
   h->origin = hs->type | (1<<27);
               ^~
```
Ref: https://github.com/curl/curl/actions/runs/8859933860/job/24330205818?pr=13489#step:5:440

+++ lib/headers.c
@@ -231,7 +231,7 @@ static CURLcode unfold_value(struct Curl_easy *data, const char *value,
   DEBUGASSERT(data->state.prevhead);
   hs = data->state.prevhead;
   olen = strlen(hs->value);
-  offset = hs->value - hs->buffer;
+  offset = (size_t)(hs->value - hs->buffer);
   oalloc = olen + offset + 1;
 
   /* skip all trailing space letters */
@@ -295,7 +295,7 @@ CURLcode Curl_headers_push(struct Curl_easy *data, const char *header,
       /* neither CR nor LF as terminator is not a valid header */
       return CURLE_WEIRD_SERVER_REPLY;
   }
-  hlen = end - header;
+  hlen = (size_t)(end - header);
 
   if((header[0] == ' ') || (header[0] == '\t')) {
     if(data->state.prevhead)
